Added random initialization, moved stuff to utils section
